requirements.txt
selenium
pandas
pytesseract
Pillow
webdriver-manager
numpy
opencv-python